% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acled_ts.R
\name{get_acled_ts}
\alias{get_acled_ts}
\title{Tidy ACLED Data into a time series}
\usage{
get_acled_ts(data_file_path)
}
\arguments{
\item{file}{path for ACLED data}
}
\value{
tstibble with columns for Quarter, EVENT_TYPE, SUB_EVENT_TYPE, COUNTRY, ACTOR1
}
\description{
Tidy ACLED Data into a time series
}
